BUG: accumulated fp error in BUB P-vector check

+++ pyentropy/statk/wrap.pyx
@@ -241,7 +241,7 @@ def bub_entropy(np.ndarray[np.float_t, ndim=1] P, int N, int dim,
     """
     if P.size != dim:
         raise ValueError, "P vector must be of length dime"
-    if np.abs(P.sum()-1.0) > np.finfo(np.float).eps:
+    if np.abs(P.sum(dtype=np.float128)-np.float128(1.0)) > np.finfo(np.float64).eps:
         raise ValueError, "sum(P) must equal 1"
     if bub_lambda_0 < 0.0:
         raise ValueError, "bub_lambda_0 must be non-negative"
